Add files via upload
ShoppingList is designed to make grocery shopping easier and quicker to access. The app will give users option to add/remove items by checking the box on/off. Friendly user interaction is a main purpose of this mobile application. I want everyone can able to approach this app with the minimum hassles.